requirements.txt
langgraph==0.2.38
langchain==0.3.5
langchain-community==0.3.3
pydantic
python-dotenv==1.0.1
streamlit>=1.28.0
reportlab>=4.0.0
Pillow>=10.0.0
